Seperation complete
