+++ SIT.Manager.Avalonia/Views/Installation/ConfigureServerView.axaml
@@ -2,7 +2,98 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+			 xmlns:ui="using:FluentAvalonia.UI.Controls"
+			 xmlns:controls="clr-namespace:SIT.Manager.Avalonia.Controls"
+			 xmlns:vm="clr-namespace:SIT.Manager.Avalonia.ViewModels.Installation"			 
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="SIT.Manager.Avalonia.Views.Installation.ConfigureServerView">
-  Welcome to Avalonia!
+             x:Class="SIT.Manager.Avalonia.Views.Installation.ConfigureServerView"
+			 x:DataType="vm:ConfigureServerViewModel">
+	<UserControl.Styles>
+		<Style Selector="ui|SymbolIcon">
+			<Setter Property="Margin" Value="0,0,8,0"/>
+		</Style>
+	</UserControl.Styles>
+
+	<Grid RowDefinitions="*, Auto">
+		<StackPanel Grid.Row="0">
+			<StackPanel>
+				<TextBlock Text="Version Selection" Classes="FrameHeading"/>
+				<Border Classes="StandardFrame">
+					<StackPanel>
+						<StackPanel Orientation="Horizontal"
+									IsVisible="{Binding !IsLoading}">
+							<TextBlock Text="Version to install:"
+									   VerticalAlignment="Center"
+									   Margin="0,0,8,0"/>
+							<ComboBox Name="VersionSelectionBox"
+									  VerticalAlignment="Center"
+									  Margin="0,0,8,0"
+									  SelectedItem="{Binding SelectedVersion}"
+									  ItemsSource="{Binding AvailableVersions}">
+								<ComboBox.ItemTemplate>
+									<DataTemplate>
+										<TextBlock Text="{Binding tag_name}"/>
+									</DataTemplate>
+								</ComboBox.ItemTemplate>
+							</ComboBox>
+						</StackPanel>
+
+						<StackPanel Margin="0,4,0,0"
+									IsVisible="{Binding !IsLoading}">
+							<TextBlock Text="No SPT-AKI install path specified"
+									   Classes="WarningMessage"/>
+							<StackPanel Orientation="Horizontal" Margin="0,8,0,0">
+								<StackPanel Orientation="Horizontal" >
+									<TextBlock Text="{DynamicResource SettingsSPTAKIPathTitle}"  VerticalAlignment="Center"/>
+									<TextBox Watermark="{DynamicResource SettingsSPTAKIInstallPathPlaceholder}"
+											 Margin="5"
+											 IsReadOnly="True"/>
+									<Button Name="ChangeInstallButton"
+											Content="{DynamicResource SettingsEFTSettingsButtonChangeTitle}"
+											ToolTip.Tip="{DynamicResource SettingsEFTChangeInstallPathToolTip}"
+											Command="{Binding ChangeServerInstallLocationCommand}"/>
+								</StackPanel>
+							</StackPanel>
+						</StackPanel>
+
+						<controls:LoadingSpinner Width="64"
+												 Height="{Binding $self.Width}"
+												 Margin="0"
+												 HorizontalAlignment="Center"
+												 VerticalAlignment="Center"
+												 IsVisible="{Binding IsLoading}"
+												 Foreground="{DynamicResource TextControlForeground}"
+												 StrokeWidth="8"/>
+					</StackPanel>
+				</Border>
+			</StackPanel>
+		</StackPanel>
+
+		<Grid Grid.Row="1"
+			  ColumnDefinitions="*, *"
+			  Margin="4,4,4,8">
+			<Button Grid.Column="0"
+					Height="50"
+					HorizontalAlignment="Center"
+					HorizontalContentAlignment="Stretch"
+					Command="{Binding BackCommand}">
+				<StackPanel Orientation="Horizontal">
+					<ui:SymbolIcon Symbol="Back"/>
+					<TextBlock Text="Back"/>
+				</StackPanel>
+			</Button>
+
+			<Button Grid.Column="1"
+					Height="50"
+					HorizontalAlignment="Center"
+					HorizontalContentAlignment="Stretch"
+					IsEnabled="{Binding IsConfigurationValid}"
+					Command="{Binding StartCommand}">
+				<StackPanel Orientation="Horizontal">
+					<ui:SymbolIcon Symbol="PlayFilled"/>
+					<TextBlock Text="Start Install"/>
+				</StackPanel>
+			</Button>
+		</Grid>
+	</Grid>
 </UserControl>
