+++ SIT.Manager.Avalonia/Views/Installation/ConfigureSitView.axaml.cs
@@ -1,10 +1,16 @@
-using Avalonia.Controls;
+using Avalonia.ReactiveUI;
+using Microsoft.Extensions.DependencyInjection;
+using ReactiveUI;
+using SIT.Manager.Avalonia.ViewModels.Installation;
 
 namespace SIT.Manager.Avalonia.Views.Installation;
-public partial class ConfigureSitView : UserControl
+
+public partial class ConfigureSitView : ReactiveUserControl<ConfigureSitViewModel>
 {
     public ConfigureSitView()
     {
         InitializeComponent();
+        this.DataContext = App.Current.Services.GetService<ConfigureSitViewModel>();
+        this.WhenActivated(disposables => { /* Handle view activation etc. */ });
     }
 }
