--- SIT.Manager.Avalonia/Views/Installation/ConfigureView.axaml
@@ -1,127 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 xmlns:ui="using:FluentAvalonia.UI.Controls"
-			 xmlns:controls="clr-namespace:SIT.Manager.Avalonia.Controls"
-			 xmlns:vm="clr-namespace:SIT.Manager.Avalonia.ViewModels.Installation"
-			 mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="SIT.Manager.Avalonia.Views.Installation.ConfigureView"
-			 x:DataType="vm:ConfigureViewModel">
-	<UserControl.Styles>
-		<Style Selector="ui|SymbolIcon">
-			<Setter Property="Margin" Value="0,0,8,0"/>
-		</Style>
-	</UserControl.Styles>
-	
-	<Grid RowDefinitions="*, Auto">
-		<StackPanel Grid.Row="0">
-			<StackPanel>
-				<TextBlock Text="Version Selection" Classes="FrameHeading"/>
-				<Border Classes="StandardFrame">
-					<StackPanel>
-						<StackPanel Orientation="Horizontal"
-									IsVisible="{Binding !IsLoading}">
-							<TextBlock Text="Version to install:"
-									   VerticalAlignment="Center"
-									   Margin="0,0,8,0"/>
-							<ComboBox Name="VersionSelectionBox"
-									  VerticalAlignment="Center"
-									  Margin="0,0,8,0"
-									  SelectedItem="{Binding SelectedVersion}"
-									  ItemsSource="{Binding AvailableVersions}">
-								<ComboBox.ItemTemplate>
-									<DataTemplate>
-										<TextBlock Text="{Binding tag_name}"/>
-									</DataTemplate>
-								</ComboBox.ItemTemplate>
-							</ComboBox>
-						</StackPanel>
-
-						<StackPanel Orientation="Horizontal"
-									Margin="0,4,0,0">
-							<StackPanel.IsVisible>
-								<MultiBinding Converter="{x:Static BoolConverters.And}">
-									<Binding Path="IsSitInstall"/>
-									<Binding Path="!IsLoading"/>
-								</MultiBinding>
-							</StackPanel.IsVisible>
-
-							<TextBlock Text="Select the mirror to use: "
-									   VerticalAlignment="Center"
-									   Margin="0,0,8,0"/>
-							<ComboBox Name="MirrorSelectionBox"
-									  Margin="0,0,8,0"
-									  VerticalAlignment="Center"
-									  SelectedItem="{Binding SelectedMirror}"
-									  ItemsSource="{Binding AvailableMirrors}">
-								<ComboBox.ItemTemplate>
-									<DataTemplate>
-										<TextBlock Text="{Binding Key}"/>
-									</DataTemplate>
-								</ComboBox.ItemTemplate>
-							</ComboBox>
-						</StackPanel>
-
-						<StackPanel Margin="0,4,0,0">
-							<StackPanel.IsVisible>
-								<MultiBinding Converter="{x:Static BoolConverters.And}">
-									<Binding Path="IsSitInstall"/>
-									<Binding Path="CurrentInstallProcessState.UsingBsgInstallPath"/>
-									<Binding Path="!IsLoading"/>
-								</MultiBinding>
-							</StackPanel.IsVisible>
-							
-							<TextBlock Text="Currently using the BSG EFT install location please install EFT to a different location to avoid any possible issues with SIT"
-									   Classes="WarningMessage"/>
-							<StackPanel Orientation="Horizontal" Margin="0,8,0,0">
-								<StackPanel Orientation="Horizontal" >
-									<TextBlock Text="{DynamicResource SettingsEFTInstallPathTitle}" Margin="0,0,8,0" VerticalAlignment="Center"/>
-									<TextBox Watermark="{DynamicResource SettingsEFTInstallPathPlaceholder}"
-											 Margin="0,0,8,0"
-											 IsReadOnly="True"/>
-									<Button Name="ChangeInstallButton"
-											Content="{DynamicResource SettingsEFTSettingsButtonChangeTitle}"
-											ToolTip.Tip="{DynamicResource SettingsEFTChangeInstallPathToolTip}"
-											Command="{Binding ChangeEftInstallLocationCommand}"/>
-								</StackPanel>
-							</StackPanel>
-						</StackPanel>
-
-						<controls:LoadingSpinner Width="64"
-												 Height="{Binding $self.Width}"
-												 Margin="0"
-												 HorizontalAlignment="Center"
-												 VerticalAlignment="Center"
-												 IsVisible="{Binding IsLoading}"
-												 Foreground="{DynamicResource TextControlForeground}"
-												 StrokeWidth="8"/>
-					</StackPanel>
-				</Border>
-			</StackPanel>
-
-			<StackPanel IsVisible="{Binding IsSitInstall}">
-				<TextBlock Text="Advanced Options" Classes="FrameHeading"/>
-				<Border Classes="StandardFrame">
-					<StackPanel>
-						<TextBlock Text="Ram Setup Goes here"/>
-						<TextBlock Text="Mods Selection Goes Here"/>
-					</StackPanel>
-				</Border>
-			</StackPanel>
-		</StackPanel>
-
-		<Button Grid.Row="1" 
-				Margin="4,4,4,8"
-				Height="50"
-				HorizontalAlignment="Center"
-				HorizontalContentAlignment="Stretch"
-				IsEnabled="{Binding IsConfigurationValid}"
-				Command="{Binding StartCommand}">
-			<StackPanel Orientation="Horizontal">
-				<ui:SymbolIcon Symbol="PlayFilled"/>
-				<TextBlock Text="Start Install"/>
-			</StackPanel>
-		</Button>
-	</Grid>
-</UserControl>
